taking decoder out of reader, too large
